+++ packages/ozone/src/api/communication/updateTemplate.ts
@@ -1,6 +1,7 @@
 import { AuthRequiredError, InvalidRequestError } from '@atproto/xrpc-server'
 import { Server } from '../../lexicon'
 import AppContext from '../../context'
+import { isDuplicateTemplateNameError } from '../../communication-service/util'
 
 export default function (server: Server, ctx: AppContext) {
   server.tools.ozone.communication.updateTemplate({
@@ -26,14 +27,24 @@ export default function (server: Server, ctx: AppContext) {
       }
 
       const communicationTemplate = ctx.communicationTemplateService(db)
-      const updatedTemplate = await communicationTemplate.update(Number(id), {
-        ...template,
-        lastUpdatedBy: updatedBy,
-      })
+      try {
+        const updatedTemplate = await communicationTemplate.update(Number(id), {
+          ...template,
+          lastUpdatedBy: updatedBy,
+        })
 
-      return {
-        encoding: 'application/json',
-        body: communicationTemplate.view(updatedTemplate),
+        return {
+          encoding: 'application/json',
+          body: communicationTemplate.view(updatedTemplate),
+        }
+      } catch (err) {
+        if (isDuplicateTemplateNameError(err)) {
+          throw new InvalidRequestError(
+            `${template.name} already exists. Please choose a different name.`,
+            'DuplicateTemplateName',
+          )
+        }
+        throw err
       }
     },
   })